# Flask/flask_base/api_documentations/go_api_users/swagger.json
{
    "schemes": [
        "http"
      ],
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "User API",
      "description": "API for managing user information"
    },
    "basePath": "/",
    "paths": {
      "/users": {
        "get": {
          "summary": "Get a list of all users",
          "responses": {
            "200": {
              "description": "Successful response",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/User"
                }
              }
            }
          }
        },
        "post": {
          "summary": "Create a new user",
          "parameters": [
            {
              "name": "user",
              "in": "body",
              "description": "User object",
              "required": true,
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "User created successfully"
            }
          }
        }
      },
      "/users/{userId}": {
        "get": {
          "summary": "Get user by ID",
          "parameters": [
            {
              "name": "userId",
              "in": "path",
              "type": "integer",
              "required": true,
              "description": "ID of the user"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "schema": {
                "$ref": "#/definitions/User"
              }
            },
            "404": {
              "description": "User not found"
            }
          }
        },
        "put": {
          "summary": "Update user by ID",
          "parameters": [
            {
              "name": "userId",
              "in": "path",
              "type": "integer",
              "required": true,
              "description": "ID of the user"
            },
            {
              "name": "user",
              "in": "body",
              "description": "Updated user object",
              "required": true,
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "User updated successfully"
            },
            "404": {
              "description": "User not found"
            }
          }
        },
        "delete": {
          "summary": "Delete user by ID",
          "parameters": [
            {
              "name": "userId",
              "in": "path",
              "type": "integer",
              "required": true,
              "description": "ID of the user"
            }
          ],
          "responses": {
            "204": {
              "description": "User deleted successfully"
            },
            "404": {
              "description": "User not found"
            }
          }
        }
      }
    },
    "definitions": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "musicliked": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  }
  